New message history handling to support Skype 6+
Skype 6 (and higher) has introduced a random bug where the dialog_partner value is sometimes written as NULL in the database, which meant that certain messages in the history were not being exported. Thanks to @watsonbox for reporting the bug.

Unfortunately, detection of "conference start" events was closely tied to the existence of the dialog_partner value, so to solve this seemingly "minor" bug there actually had to be some major restructurings to avoid losing the conference detection functionality.

We now start our 1on1 conversation scanning by determining the convo_id of the partner (instead of using dialog_partner). But *just* doing that would mean that all "You added X to a conference" / "X added you to a conference" events would be missed, so I also had to add some pretty intricate event scanning to the SQL statement, which now looks for ALL Type10 ("person added to conference") events where *either* you're the inviter and the 1on1 partner is in the list of invitees, *or* where your 1on1 partner is the inviter and you are in the list of invitees.

This solves the message history issue with Skype 6+, and remains completely backwards-compatible with older databases!

There is also some good news: Skype 6 and higher now always saves Type10 ("person added to conference") events, *even* when *you* were the one being added. In the past, they only saved them when you were the one inviting *someone else*. As a result of this change, we can now reliably catch "X added you to a conference" events and display those too in the history output! Not even Skype itself does this!

To celebrate the fact that even incoming conference creations are now caught, nicer language was added to the history output; instead of just showing conferences that you create, it now says "You added Steve Jobs to a group conversation [link to the conversation]" and "Steve Jobs added you to a group conversation [link]". This also reflects the change in language in Skype, where they now call them "group conversations".

Lastly, a bunch of developer test databases, test logins and a test-run script added to the source package to facilitate any future testing. Note that the automatic test-export script only runs on Macs.