Support for Skype 6.22+ cloud file transfers
This is a new feature introduced in Skype 6.22 and up, where it automatically uploads images (no other formats for now) into the cloud and then sends a reference to the recipient. It introduces an entirely new event type (201).

Skype currently does automatic cloud transfers whenever you send a .jpg or .png file, but they may add other file formats in the future since they're working hard to make Skype completely cloud-based.

Right now they only store a single file _per_ type 201 event in the database (so multiple simultaneous image transfers are stored as individual events); but in order to be as forwards-compatible as possible, the new parsing code was written with expansion in mind so that it should automatically work even if they decide to add future support for multiple cloud files per message.