New contact scanner to avoid rare Skype 6+ bug
Skype 6 and later will extremely rarely corrupt the database fields, which in some cases leads to an empty "author" field. Since we rely on that field to discover all people you've ever encountered (regardless of contact method), we have to insert an "IS NOT NULL" check to make sure that we never grab an empty name if there's database corruption.