Better handling of contacts without 1on1 chats
If there is no 1on1 chat history for a person in your Skype database, it means that you've only encountered them in a conference (most likely because a mutual friend added them to the group chat). In that case, we simply output a hidden HTML comment stating that there was no 1on1 chat history for that person. This is an improvement over the previous handling, which assumed that every person in your Skype database had a 1on1 chat and threw an error if it failed to find one.