Exported histories now show the active UTC offset
This helps things greatly when sharing exported histories internationally. If you're telling it to output in local time, it will show your local UTC offset, such as "All times are UTC+02:00" or "UTC-04:00" or "UTC+09:30" and so on. If you're telling it to output in UTC, or you're in a timezone which is identical to the current UTC time (such as Iceland with DST active), it will say "UTC+00:00". This means that regardless of output time format, the reader will know exactly what time offsets are being used.