Always display Edited state even when Pending
This update brings us in line with a minor change in Skype 6+. In the past, Skype didn't show the "Edited" icon until a message was fully delivered. If a message was still pending, it just showed the string "Pending". Since the behavior has changed, the exporter has been updated to always show the "Edited" icon if a message has been edited.

Thanks to Santosh Thoduka (@sthoduka) for discovering this new change!

Closes #6