Improved contact scanner to ignore spammers
The contact scanner has received an upgrade! It used to look for all "authors" of events in the Messages table. It was a great way of ensuring that all friends are found; even long-deleted ones. Unfortunately, this scanning method had two drawbacks. The first, minor drawback was that it only found your friends if they had sent at least one event/message to you (that's pretty much always true, so it was no big deal). But the bigger issue was that spammers will often send you "please add me as a friend!" requests, which are added into the Messages table with them as the "author" of the event, and were picked up by the contact scanner. This meant that most people running SkypeExport would be exporting a bunch of blank history files for spammers they never accepted the friend requests from.

The new and improved scanner is now a lot more picky: It finds all unique, non-null "author" (someone who sent a message) and "dialog_partner" fields (the latter finds any remaining people who you have talked to, but who have never sent messages *back* to you, which fixes the minor issue mentioned earlier). The cherry on top is that it now *ignores* all "type 50" (friend request events).

This improvement gives you a perfect friends list export, without seeing any annoying spammer contacts anymore.