Improved displayname scanner with SkypeID fallback
There has been an increasing "cloudification" of Skype ever since version 6. One of those changes is the "cloud history" feature, which means that a fresh Skype installation will pull down a partial chat history from their servers and store it in the local database.

The problem with their implementation of cloud history is that they don't transmit any displayname data for the people you've talked to. The result of their bad history implementation meant that the displayname scanner would fall back to an empty string as their name if none could be found. That problem presented itself in the form of sentences like "Chat History with  (test.acc2001)", whereas it *should* have said "Chat History with Test Acc2001 (test.acc2001)".

The displayname scanner has now been improved so that it can perform a fallback to the raw SkypeID (this is done by default). If your history database lacks a proper displayname for the person, it will simply do the next-best-thing and use their SkypeID instead, which gives you improved sentences like "Chat History with test.acc2001 (test.acc2001)".

It's only a small cosmetic tweak, but it was still important to fix this since their cloud history feature is becoming more and more pervasive with all of their cross-device history sharing and so on...