New message direction scanner to fix Skype 6+ bug
Skype's "cloud history" feature is bugged in many ways and omits a lot of database fields or incorrectly fills them in, and unfortunately it also has a serious bug regarding their chatmsg_status field... When your computer downloads the past chat history from Skype's servers and merges it into the local database, the data for the earliest events in your 1on1 chats *and* conferences will *all* have a chatmsg_status of "4" (incoming event) even if *you* sent them *outgoing*.

That obviously caused mayhem when we were trying to export those bugged events, since it would incorrectly believe that things *we* said/did were being said/done by *someone else*.

The solution was to detect the message direction using a newer Skype database field instead (it's been there for over half a decade, since at least Skype 5, probably even 4 or earlier). The name of that field is "sending_status", and it is "1" (pending; only in Skype 5 or earlier) or "2" (delivered) for all outgoing events. More importantly, its value is transferred correctly in Skype's downloaded "cloud history"!

However, the "sending_status" field doesn't tell us if a message is pending anymore on Skype 6+, so we have a hybrid solution where we *also* retrieve the legacy "chatmsg_status" field, and if it's "1" (pending; still saved that way by Skype 6+ until a message is delivered) then we know that the message is not yet delivered to the recipient. Skype changes "chatmsg_status" to "2" when the message has been delivered.

With this hybrid solution, we now have reliable event direction detection again -- regardless of whether your history data came from the cloud or locally! And most importantly: It's still backwards-compatible with all Skype database versions up to *at least* half a decade ago!