Generate valid log filenames for weird usernames
Microsoft has recently created "fake" Skype accounts for all of their
Windows Live IDs. So if your Microsoft ID was "cooldude123", it would
create a Skype account called "live:cooldude123". The colon is still an
illegal character in real Skype IDs and the only way to get it is via
this auto-creation.

The colon is also an invalid character on your filesystem, so we now
send every SkypeID through a validation process that replaces all illegal
filesystem characters with "$" instead, so the above would be output as
"live$coolude123.skypelog.htm", and will now be correctly written to
disk thanks to having a valid filename.

The reason for choosing the "$" character is that it's different from
characters that *can* appear in Skype IDs, and it is legal in filenames
both locally and on the web.

This closes issue #7